first loop trough the code

+++ missions/C/hello_int/init.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env sh
 
 cd "$GSH_HOME"
-cp "$MISSION_DIR/hello.c" "hello.c"
+if [ ! -e hello.c ];
+then
+  cp "$MISSION_DIR/hello.c" "hello.c"
+fi
